+++ ContentPatcher/Framework/TokenParser.cs
@@ -5,6 +5,7 @@ using ContentPatcher.Framework.Migrations;
 using ContentPatcher.Framework.Tokens;
 using ContentPatcher.Framework.Tokens.Json;
 using Newtonsoft.Json.Linq;
+using Pathoschild.Stardew.Common.Utilities;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework
@@ -24,6 +25,9 @@ namespace ContentPatcher.Framework
         /// <summary>The migrator which validates and migrates content pack data.</summary>
         public IMigration Migrator { get; }
 
+        /// <summary>The mod IDs which are currently installed.</summary>
+        public InvariantHashSet InstalledMods { get; }
+
 
         /*********
         ** Public methods
@@ -32,37 +36,31 @@ namespace ContentPatcher.Framework
         /// <param name="tokenContext">The tokens available for this content pack.</param>
         /// <param name="forMod">The manifest for the content pack being parsed.</param>
         /// <param name="migrator">The migrator which validates and migrates content pack data.</param>
-        public TokenParser(IContext tokenContext, IManifest forMod, IMigration migrator)
+        /// <param name="installedMods">The mod IDs which are currently installed.</param>
+        public TokenParser(IContext tokenContext, IManifest forMod, IMigration migrator, InvariantHashSet installedMods)
         {
             this.Context = tokenContext;
             this.ForMod = forMod;
             this.Migrator = migrator;
+            this.InstalledMods = installedMods;
         }
 
         /// <summary>Parse a string which can contain tokens, and validate that it's valid.</summary>
         /// <param name="rawValue">The raw string which may contain tokens.</param>
+        /// <param name="assumeModIds">Mod IDs to assume are installed for purposes of token validation.</param>
         /// <param name="error">An error phrase indicating why parsing failed (if applicable).</param>
         /// <param name="parsed">The parsed value.</param>
-        public bool TryParseStringTokens(string rawValue, out string error, out IParsedTokenString parsed)
+        public bool TryParseStringTokens(string rawValue, InvariantHashSet assumeModIds, out string error, out IParsedTokenString parsed)
         {
             // parse
             parsed = new TokenString(rawValue, this.Context);
             if (!this.Migrator.TryMigrate(parsed, out error))
                 return false;
 
-            // validate unknown tokens
-            IContextualState state = parsed.GetDiagnosticState();
-            if (state.InvalidTokens.Any())
-            {
-                error = $"found unknown tokens ({string.Join(", ", state.InvalidTokens.OrderBy(p => p))})";
-                parsed = null;
-                return false;
-            }
-
             // validate tokens
             foreach (LexTokenToken lexToken in parsed.GetTokenPlaceholders(recursive: false))
             {
-                if (!this.TryValidateToken(lexToken, out error))
+                if (!this.TryValidateToken(lexToken, assumeModIds, out error))
                     return false;
             }
 
@@ -74,9 +72,10 @@ namespace ContentPatcher.Framework
 
         /// <summary>Parse a JSON structure which can contain tokens, and validate that it's valid.</summary>
         /// <param name="rawJson">The raw JSON structure which may contain tokens.</param>
+        /// <param name="assumeModIds">Mod IDs to assume are installed for purposes of token validation.</param>
         /// <param name="error">An error phrase indicating why parsing failed (if applicable).</param>
         /// <param name="parsed">The parsed value, which may be legitimately <c>null</c> even if successful.</param>
-        public bool TryParseJsonTokens(JToken rawJson, out string error, out TokenisableJToken parsed)
+        public bool TryParseJsonTokens(JToken rawJson, InvariantHashSet assumeModIds, out string error, out TokenisableJToken parsed)
         {
             if (rawJson == null || rawJson.Type == JTokenType.Null)
             {
@@ -93,7 +92,7 @@ namespace ContentPatcher.Framework
             // validate tokens
             foreach (LexTokenToken lexToken in parsed.GetTokenStrings().SelectMany(p => p.GetTokenPlaceholders(recursive: false)))
             {
-                if (!this.TryValidateToken(lexToken, out error))
+                if (!this.TryValidateToken(lexToken, assumeModIds, out error))
                     return false;
             }
 
@@ -106,13 +105,23 @@ namespace ContentPatcher.Framework
 
         /// <summary>Validate whether a token referenced in a content pack field is valid.</summary>
         /// <param name="lexToken">The lexical token reference to check.</param>
+        /// <param name="assumeModIds">Mod IDs to assume are installed for purposes of token validation.</param>
         /// <param name="error">An error phrase indicating why validation failed (if applicable).</param>
-        public bool TryValidateToken(LexTokenToken lexToken, out string error)
+        public bool TryValidateToken(LexTokenToken lexToken, InvariantHashSet assumeModIds, out string error)
         {
             // token doesn't exist
             IToken token = this.Context.GetToken(lexToken.Name, enforceContext: false);
             if (token == null)
             {
+                // special case: requires a missing mod that's checked via HasMod
+                string requiredModId = lexToken.GetProviderModId();
+                if (assumeModIds?.Contains(requiredModId) == true && !this.InstalledMods.Contains(requiredModId))
+                {
+                    error = null;
+                    return true;
+                }
+
+                // else error
                 error = $"'{lexToken}' can't be used as a token because that token could not be found.";
                 return false;
             }
@@ -120,8 +129,11 @@ namespace ContentPatcher.Framework
             // token requires dependency
             if (token is ModProvidedToken modToken && !this.ForMod.HasDependency(modToken.Mod.UniqueID, canBeOptional: false))
             {
-                error = $"'{lexToken}' can't be used because it's provided by {modToken.Mod.Name} (ID: {modToken.Mod.UniqueID}), but {this.ForMod.Name} doesn't list it as a dependency.";
-                return false;
+                if (assumeModIds == null || !assumeModIds.Contains(modToken.Mod.UniqueID))
+                {
+                    error = $"'{lexToken}' can't be used because it's provided by {modToken.Mod.Name} (ID: {modToken.Mod.UniqueID}), but {this.ForMod.Name} doesn't list it as a dependency and the patch doesn't have an immutable {ConditionType.HasMod} condition for that mod.";
+                    return false;
+                }
             }
 
             // no error found